Added methods to interface between Android and Python.
